requirements.txt
# Core Flask Dependencies
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.6.0
Flask-CORS==4.0.0
Flask-Mail==0.9.1
Flask-Limiter==3.5.0
python-dotenv==1.0.0

# Database & ORM
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1

# Security & Authentication
bcrypt==4.1.2
cryptography>=42.0.0
PyJWT==2.8.0
email-validator==2.1.0
Flask-WTF==1.2.1

# Bitcoin & Crypto
bitcoinlib>=0.7.0
base58==2.1.1
bech32==1.2.0

# API & Serialization
marshmallow==3.20.2
webargs==8.3.0
apispec==6.3.0
apispec-webframeworks==0.5.2

# Async & Background Tasks
celery==5.3.4
redis==5.0.1
kombu==5.3.4

# HTTP & Requests
requests==2.31.0
urllib3==2.1.0

# Configuration & Environment
python-dotenv==1.0.0
click==8.1.7

# Production Server
gunicorn==21.2.0
gevent==23.9.1

# Rust-Python Binding
maturin==1.4.0

# Development Dependencies
pytest==7.4.4
pytest-cov==4.1.0
pytest-flask==1.3.0
pytest-mock==3.12.0
factory-boy==3.3.0

# Code Quality
black==23.12.0
isort==5.13.2
flake8==6.1.0
mypy==1.7.1
pre-commit==3.6.0

# Documentation
sphinx==7.2.6
sphinx-rtd-theme==2.0.0

# Monitoring & Logging
structlog==23.2.0
sentry-sdk[flask]==1.39.2
